more cyclic dependencies cleanup
