+++ src/format/html/format-html-constants.ts
@@ -1,0 +1,7 @@
+/*
+ * format-html-constants.ts
+ *
+ * Copyright (C) 2020-2023 Posit Software, PBC
+ */
+
+export const kQuartoHtmlDependency = "quarto-html";
