+++ src/format/html/format-html-types.ts
@@ -1,0 +1,28 @@
+/*
+ * format-html-types.ts
+ *
+ * Copyright (C) 2020-2023 Posit Software, PBC
+ */
+
+import { Metadata } from "../../config/types.ts";
+
+export interface HtmlFormatFeatureDefaults {
+  tabby?: boolean;
+  copyCode?: boolean;
+  anchors?: boolean;
+  hoverCitations?: boolean;
+  hoverFootnotes?: boolean;
+  figResponsive?: boolean;
+  codeAnnotations?: boolean;
+}
+
+export interface HtmlFormatTippyOptions {
+  theme?: string;
+  parent?: string;
+  config?: Metadata;
+}
+
+export interface HtmlFormatScssOptions {
+  quartoBase?: boolean;
+  quartoCssVars?: boolean;
+}
