+++ src/format/html/format-html.ts
@@ -1,9 +1,8 @@
 /*
-* format-html.ts
-*
-* Copyright (C) 2020-2022 Posit Software, PBC
-*
-*/
+ * format-html.ts
+ *
+ * Copyright (C) 2020-2022 Posit Software, PBC
+ */
 import { join } from "path/mod.ts";
 import { warning } from "log/mod.ts";
 
@@ -100,6 +99,12 @@ import {
 } from "./format-html-notebook.ts";
 import { ProjectContext } from "../../project/types.ts";
 import { kListing } from "../../project/types/website/listing/website-listing-shared.ts";
+import {
+  HtmlFormatFeatureDefaults,
+  HtmlFormatScssOptions,
+  HtmlFormatTippyOptions,
+} from "./format-html-types.ts";
+import { kQuartoHtmlDependency } from "./format-html-constants.ts";
 
 export function htmlFormat(
   figwidth: number,
@@ -153,29 +158,6 @@ export function htmlFormat(
       },
     },
   );
-}
-
-export const kQuartoHtmlDependency = "quarto-html";
-
-export interface HtmlFormatFeatureDefaults {
-  tabby?: boolean;
-  copyCode?: boolean;
-  anchors?: boolean;
-  hoverCitations?: boolean;
-  hoverFootnotes?: boolean;
-  figResponsive?: boolean;
-  codeAnnotations?: boolean;
-}
-
-export interface HtmlFormatTippyOptions {
-  theme?: string;
-  parent?: string;
-  config?: Metadata;
-}
-
-export interface HtmlFormatScssOptions {
-  quartoBase?: boolean;
-  quartoCssVars?: boolean;
 }
 
 export async function htmlFormatExtras(
@@ -707,13 +689,13 @@ function htmlFormatPostprocessor(
     );
 
     for (let i = 0; i < tables.length; ++i) {
-      const table = (tables[i] as Element);
+      const table = tables[i] as Element;
       if (table.getAttribute("data-quarto-disable-processing")) {
         continue;
       }
       table.removeAttribute("data-quarto-postprocess-tables");
       table.querySelectorAll("tr").forEach((tr) => {
-        const { children } = (tr as Element);
+        const { children } = tr as Element;
         for (let j = 0; j < children.length; ++j) {
           const child = children[j] as Element;
           if (child.tagName === "TH" || child.tagName === "TD") {
